BEGIN: prototyping new storage functions
Since Firefox no longer allows script level access to the localStorage
object, this version of Enhanced Steam will attempt to use the IndexedDB
API instead.

These functions should correctly create the database, and should
successfully emulate the setValue, getValue, and delValue functions of
the Chrome version.  This should allow for smooth code portability.

Logging messages are staying in the code for now until any and all
issues are worked out.  This assists with easier debugging.

The overall goal here is to emulate these functions so that data can be
cached locally to reduce the number of API requests.